refactored moving of backend files
