+++ src/app/webApp.js
@@ -4,15 +4,15 @@ var path = require('path');
 var http = require('http');
 var socketio = require('socket.io');
 var bodyParser = require('body-parser');
-var {webPort, servicePort, hostAddress} = require('../common/config');
-var alucardLogger = require('./alucardLogger');
-var romSearchService = require('./romSearchService');
-var jobService = require('./jobService');
-var postRequest = require('./webClient').postRequest;
-var serviceMessageTypes = require('./serviceMessageTypes');
-var progressType = require('./progressTypes');
-var pingPi = require('./piMonitor').pingPi;
-var piStatusTypes = require('./piStatusTypes');
+var {servicePort, hostAddress} = require('../common/config');
+var alucardLogger = require('../common/logging/alucardLogger');
+var romSearchService = require('../web/webAppServices/romSearchService');
+var jobService = require('../web/webAppServices/jobService');
+var {postRequest} = require('../common/webService/webClient');
+var serviceMessageTypes = require('../common/enums/serviceMessageTypes');
+var progressType = require('../common/enums/progressTypes');
+var piStatusTypes = require('../common/enums/piStatusTypes');
+var {pingPi}= require('../fileService/fileTransfer/piMonitor');
 
 app.use(bodyParser.json());
 app.use(express.static(path.resolve(__dirname, '../static')));
