+++ src/common/logging/fileLogger.js
@@ -1,8 +1,8 @@
 var fs = require('fs');
-var {baseOutputDir, logFileDir} = require('../common/config');
+var {baseOutputDir, logFileDir} = require('../../common/config');
 var path = require('path');
 var levels = require('./logConfig').levels;
-var ensureDirectory = require('../utility/fileSystemHelpers').ensureDirectory;
+var {ensureDirectory} = require('../utility/fileSystemHelpers');
 
 function errLog(err) {
     console.log(err);
