+++ src/common/models/mongoosePromisified.js
@@ -1,6 +1,6 @@
 var mongoose = require('mongoose');
-var Promise = require('q').Promise;
-var {dbConfig: {dbAddress, dbName}} = require('../common/config');
+var {Promise} = require('q');
+var {dbConfig: {dbAddress, dbName}} = require('../../common/config');
 
 mongoose.Promise = Promise;
 //TODO only connect on construction of model
